requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.14
django-appconf==1.0.6
django-compressor==4.5.1
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.7
MarkupSafe==2.1.5
mysqlclient==2.2.0
packaging==24.1
Pillow==10.0.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
rcssmin==1.1.2
referencing==0.35.1
requests==2.32.3
rjsmin==1.2.2
rpds-py==0.20.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zipp==3.20.2
